Add engine parameter to Forward/BackwardViewMigration. Generate multiple migration operations for each defined engine.

+++ django_db_views/migration_functions.py
@@ -1,9 +1,12 @@
+from django.conf import settings
 from django.utils.deconstruct import deconstructible
 
 
 class ViewMigration(object):
 
-    def __init__(self, view_definition: str, table_name: str):
+    def __init__(self, view_definition: str, table_name: str, engine=None):
+        self.view_engine = engine
+        self.db_engine = settings.DATABASES['default']['ENGINE']
         self.view_definition = view_definition
         self.table_name = table_name
 
@@ -12,7 +15,7 @@ class ViewMigration(object):
 class ForwardViewMigration(ViewMigration):
 
     def __call__(self, apps, schema_editor):
-        if self.view_definition:
+        if self.view_definition and self.db_engine == self.view_engine:
             schema_editor.execute("DROP VIEW IF EXISTS %s;" % self.table_name)
             schema_editor.execute("CREATE VIEW %s as %s;" % (self.table_name, self.view_definition))
 
@@ -21,6 +24,7 @@ class ForwardViewMigration(ViewMigration):
 class BackwardViewMigration(ViewMigration):
 
     def __call__(self, apps, schema_editor):
-        schema_editor.execute("DROP VIEW IF EXISTS %s;" % self.table_name)
-        if self.view_definition:
+        if self.db_engine == self.view_engine:
+            schema_editor.execute("DROP VIEW IF EXISTS %s;" % self.table_name)
+        if self.view_definition and self.db_engine == self.view_engine:
             schema_editor.execute("CREATE VIEW %s as %s;" % (self.table_name, self.view_definition))
